+++ NEWS.md
@@ -6,8 +6,10 @@
 * Fix bcbioSmallSizeDist to avoid duplication of columns when getting adapter information
 
 ## Feature
-* Add sequence matrix for cluster analysis.
-* Plot size disribution by small RNA type.
+* Re-organize cluster figure in QC report
+* Use multiqc file to read metrics
+* Add sequence matrix for cluster analysis
+* Plot size disribution by small RNA type
 * Change to use varianzeStabilization from DESeq2 to normalize counts,
   as consequence, name of slot of normalized count changed to `log`
 * Load final folder using project folder path instead